backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
opencv-python==4.8.1.78
mediapipe==0.10.21
numpy==1.26.2
pyjwt==2.8.0
mysql-connector-python==9.1.0
